Refactor presence
